Various changes
Created scss modules for each file.
Changed nextjs Images to regular image tags.
Added import for each scss module even if not used yet.
Added option to add descriptions to images.
