+++ src/components/atoms/Icon/Icon.tsx
@@ -1,5 +1,7 @@
 import Image from "next/image";
 
+import styles from "./Icon.module.scss";
+
 type IconProps = {
   src: string;
   alt: string;
