+++ src/components/molecules/GalleryImage/GalleryImage.module.scss
@@ -1,0 +1,3 @@
+.GalleryImage {
+  display: inline-block;
+}
