+++ src/components/organisms/Gallery/Gallery.module.scss
@@ -7,5 +7,8 @@
   img:hover {
     border: solid 3px white;
     scale: 1.2;
+    p {
+      font-size: large;
+    }
   }
 }
