+++ src/components/organisms/Gallery/Gallery.tsx
@@ -1,4 +1,4 @@
-import GalleryImage from "@/components/molecules/GalleryCard/GalleryCard";
+import GalleryImage from "@/components/molecules/GalleryImage/GalleryImage";
 
 import styles from "./Gallery.module.scss";
 
@@ -6,14 +6,14 @@ const Gallery = () => {
   return (
     <div className={styles.list}>
       <GalleryImage src={`/a.jpg`} description={"test"} />
-      <GalleryImage src={`/b.jpg`} description={""} />
+      <GalleryImage src={`/b.jpg`} />
       <GalleryImage src={`/c.jpg`} description={"test3"} />
-      <GalleryImage src={`/d.jpg`} description={""} />
-      <GalleryImage src={`/e.jpg`} description={""} />
-      <GalleryImage src={`/f.jpg`} description={""} />
+      <GalleryImage src={`/d.jpg`} />
+      <GalleryImage src={`/e.jpg`} />
+      <GalleryImage src={`/f.jpg`} />
       <GalleryImage src={`/g.jpg`} description={"salut ma numesc fotograful"} />
-      <GalleryImage src={`/h.jpg`} description={""} />
-      <GalleryImage src={`/i.jpg`} description={""} />
+      <GalleryImage src={`/h.jpg`} />
+      <GalleryImage src={`/i.jpg`} />
     </div>
   );
 };
